requirements.txt
hydra-core==1.3.1
torch==1.11.0
numpy==1.21.6
torchvision==0.12.0